+++ packages/@tenoxui-preset-tailwind/src/lib/property/svg.ts
@@ -5,13 +5,13 @@ import { is } from 'cssrxp'
 
 export const svg: Property = {
   fill: ({ key = '', value = '', unit = '', secondValue = '', secondUnit = '' }) =>
-    !value || key || secondUnit || !is.color.test(value)
+    !value || key || secondUnit || (!is.color.test(value) && value !== 'current')
       ? null
-      : createColorType('fill', value, secondUnit),
+      : createColorType('fill', value, secondValue),
   stroke: ({ key = '', value = '', unit = '', secondValue = '', secondUnit = '' }) =>
     !value || key || secondUnit
       ? null
-      : is.color.test(value)
-        ? createColorType('stroke', value, secondUnit)
+      : is.color.test(value) || value === 'current'
+        ? createColorType('stroke', value, secondValue)
         : `value:${toKebab('strokeWidth')}: ${value + unit}`
 }